requirements.txt
django==1.7.4
django_extensions==1.5.0
Werkzeug==0.10.1
gunicorn==19.2.1
raven==5.1.1
psycopg2==2.6
fabric==1.10.1
django-uuidfield==0.5.0
djangorestframework==3.0.5
pygments==2.0.2
git+https://github.com/bfirsh/django-ordered-model.git#egg=django-ordered-model
mock==1.0.1
